chore: use header tags more often and at right places
