+++ packages/frontend/src/components/Strategies/Crab/StrategyPerformance.tsx
@@ -21,10 +21,10 @@ import {
 } from '@state/crab/atoms'
 import { BIG_ZERO, CRABV2_START_DATE } from '@constants/index'
 import { useETHPrice } from '@hooks/useETHPrice'
-import { useOSQTHPrice } from '@hooks/useOSQTHPrice'
 import { formatCurrency, formatNumber } from '@utils/formatter'
 import { pnlGraphOptions } from '@constants/diagram'
 import useAppMemo from '@hooks/useAppMemo'
+import PerformanceMetric from '@components/Strategies/PerformanceMetric'
 
 const useTextFieldStyles = makeStyles((theme) =>
   createStyles({
@@ -73,29 +73,6 @@ const CustomTextField: React.FC<TextFieldProps> = ({ inputRef, label, InputProps
         }}
         {...props}
       />
-    </Box>
-  )
-}
-
-const PerformanceMetric: React.FC<{ label: string; value: number }> = ({ label, value }) => {
-  const classes = useStyles()
-
-  return (
-    <Box display="flex" justifyContent="flex-end" gridGap="6px">
-      <Typography className={classes.textSmall}>{label}</Typography>
-
-      <Box minWidth="6ch" display="flex" justifyContent="flex-end">
-        <Typography
-          className={clsx(
-            classes.textSmall,
-            classes.textMonospace,
-            value >= 0 ? classes.colorSuccess : classes.colorError,
-          )}
-        >
-          {value >= 0 && '+'}
-          {formatNumber(value)}%
-        </Typography>
-      </Box>
     </Box>
   )
 }
@@ -187,7 +164,9 @@ const StrategyPerformance: React.FC<StrategyPerformanceProps> = ({ strategyPnLSe
         </Typography>
 
         <Box display="flex" alignItems="baseline" gridGap="12px">
-          <Typography className={classes.description}>Annualized USDC Return</Typography>
+          <Typography variant="h4" className={classes.description}>
+            Annualized USDC Return
+          </Typography>
 
           <Box position="relative" top="3px">
             <Tooltip title={<TooltipTitle />}>
@@ -199,7 +178,9 @@ const StrategyPerformance: React.FC<StrategyPerformanceProps> = ({ strategyPnLSe
 
       <Box display="flex" gridGap="12px">
         <Typography className={clsx(classes.description, classes.textMonospace)}>{formatCurrency(tvl, 0)}</Typography>
-        <Typography className={classes.description}>Open Interest</Typography>
+        <Typography variant="h4" className={classes.description}>
+          Open Interest
+        </Typography>
       </Box>
 
       <Box display="flex" justifyContent="space-between" alignItems="flex-end" gridGap="12px" flexWrap="wrap">
